refactor : rename product to product_s
