+++ package.json
@@ -40,8 +40,10 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-apexcharts": "^1.7.0",
+    "react-bootstrap": "^2.10.6",
     "react-copy-to-clipboard": "^5.1.0",
     "react-datepicker": "^7.5.0",
+    "react-daum-postcode": "^3.1.3",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
     "react-icons": "^5.4.0",
@@ -59,8 +61,7 @@
     "vite": "^5.2.10",
     "vite-jsconfig-paths": "^2.0.1",
     "web-vitals": "^3.5.2",
-    "yup": "^1.4.0",
-    "react-daum-postcode": "^3.1.3"
+    "yup": "^1.4.0"
   },
   "eslintConfig": {
     "extends": [
