+++ src/pages/acorn-pages/product/Product.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Routes, Route } from "react-router-dom";
 import ProductBList from './productB/ProductBList';
 import PRODUCT_B from './PRODUCT_B';
-import PRODUCT from './PRODUCTS';
+import PRODUCT_S from './PRODUCT_S';
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
       <Routes>
           <Route path="/" element={<PRODUCT_B />}></Route>
           <Route path="/productB" element={<ProductBList />}></Route>
-          <Route path="/product" element={<PRODUCT />}></Route>
+          <Route path="/productS" element={<PRODUCT_S />}></Route>
       </Routes>
     </>
   );
